Continuação da criação do CRUD de produção

+++ Producao/Models/MateriaPrimaCheckboxViewModel.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Nome { get; set; }
         public bool Selecionado { get; set; }
-        public double Quantidade { get; set; }
+        public int Quantidade { get; set; }
 
     }
 }
